requirements.txt
# Core dependencies
openai>=1.0.0
google-genai>=1.0.0
chromadb>=0.4.0
rich>=13.0.0
click>=8.0.0
python-dotenv>=1.0.0
watchdog>=3.0.0
sentence-transformers>=2.2.0
pyyaml>=6.0.0
requests>=2.28.0
tiktoken>=0.5.0           # Token counting for cost estimates
flask>=3.0.0              # Web interface
flask-cors>=4.0.0         # CORS support for API

# Optional AI providers (install as needed)
anthropic>=0.18.0       # For Claude models
groq>=0.4.0             # For Groq fast inference
ollama>=0.1.0           # For local Ollama models

# Optional file format support (install as needed)
pypdf>=4.0.0            # For PDF files
python-docx>=1.0.0      # For Word documents (.docx)
striprtf>=0.0.26        # For RTF files
